pyproject.toml
[tool.poetry]
name = "school"
version = "0.1.0"
description = ""
authors = ["guilherme <guile.hm@hotmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.3"
flake8 = "^3.9.2"
isort = "^5.8.0"
uvicorn = {extras = ["standard"], version = "^0.13.4"}
dj-database-url = "^0.5.0"
psycopg2-binary = "^2.8.6"
whitenoise = "^5.2.0"
django-extensions = "^3.1.3"
django-debug-toolbar = "^3.2.1"
pytest-django = "^4.3.0"
graphene-django = "^2.15.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
